requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Session==0.8.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2025.2
qrcode==8.2
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
